Remove some more redundant events
